refactor(architecture): remove final runtime cycle
